# Add BSF and BSR fallbacks for BitOperations methods