--- a/src/libraries/System.Private.CoreLib/src/System/Numerics/BitOperations.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Numerics/BitOperations.cs
@@ -67,10 +67,18 @@ namespace System.Numerics
                 return ArmBase.LeadingZeroCount(value);
             }
 
-            // Unguarded fallback contract is 0->31
+            // Unguarded fallback contract is 0->31, BSR contract is 0->undefined
             if (value == 0)
             {
                 return 32;
+            }
+
+            if (X86Base.IsSupported)
+            {
+                // LZCNT returns index starting from MSB, whereas BSR gives the index from LSB.
+                // The XOR here is equivalent to subtraction since the BSR result is always between 0 and 31.
+                // This saves one instruction, as subtraction from constant requires either MOV/SUB or NEG/ADD.
+                return 31 ^ (int)X86Base.BitScanReverse(value);
             }
 
             return 31 - Log2SoftwareFallback(value);
@@ -94,6 +102,12 @@ namespace System.Numerics
             if (ArmBase.Arm64.IsSupported)
             {
                 return ArmBase.Arm64.LeadingZeroCount(value);
+            }
+
+            if (X86Base.X64.IsSupported)
+            {
+                // BSR contract is 0->undefined
+                return value == 0 ? 64 : 63 ^ (int)X86Base.X64.BitScanReverse(value);
             }
 
             uint hi = (uint)(value >> 32);
@@ -130,13 +144,19 @@ namespace System.Numerics
             // 1000..    0      31-0    31
             if (Lzcnt.IsSupported)
             {
-                // LZCNT contract is 0->32
-                return 31 - (int)Lzcnt.LeadingZeroCount(value);
+                return 31 ^ (int)Lzcnt.LeadingZeroCount(value);
             }
 
             if (ArmBase.IsSupported)
             {
                 return 31 - ArmBase.LeadingZeroCount(value);
+            }
+
+            // BSR returns the answer we're looking for directly.
+            // However BSR is much slower than TZCNT on AMD processors, so we leave it as a fallback only.
+            if (X86Base.IsSupported)
+            {
+                return (int)X86Base.BitScanReverse(value);
             }
 
             // Fallback contract is 0->0
@@ -160,13 +180,17 @@ namespace System.Numerics
 
             if (Lzcnt.X64.IsSupported)
             {
-                // LZCNT contract is 0->64
-                return 63 - (int)Lzcnt.X64.LeadingZeroCount(value);
+                return 63 ^ (int)Lzcnt.X64.LeadingZeroCount(value);
             }
 
             if (ArmBase.Arm64.IsSupported)
             {
                 return 63 - ArmBase.Arm64.LeadingZeroCount(value);
+            }
+
+            if (X86Base.IsSupported)
+            {
+                return (int)X86Base.X64.BitScanReverse(value);
             }
 
             uint hi = (uint)(value >> 32);
@@ -301,10 +325,15 @@ namespace System.Numerics
                 return ArmBase.LeadingZeroCount(ArmBase.ReverseElementBits(value));
             }
 
-            // Unguarded fallback contract is 0->0
+            // Unguarded fallback contract is 0->0, BSF contract is 0->undefined
             if (value == 0)
             {
                 return 32;
+            }
+
+            if (X86Base.IsSupported)
+            {
+                return (int)X86Base.BitScanForward(value);
             }
 
             // uint.MaxValue >> 27 is always in range [0 - 31] so we use Unsafe.AddByteOffset to avoid bounds check
@@ -343,6 +372,13 @@ namespace System.Numerics
             {
                 return ArmBase.Arm64.LeadingZeroCount(ArmBase.Arm64.ReverseElementBits(value));
             }
+
+            if (X86Base.X64.IsSupported)
+            {
+                // BSF contract is 0->undefined
+                return value == 0 ? 64 : (int)X86Base.X64.BitScanForward(value);
+            }
+
             uint lo = (uint)value;
 
             if (lo == 0)