… src/coreclr/src/inc/readytoruninstructionset.h
… src/coreclr/src/jit/hwintrinsiclistxarch.h
… src/coreclr/src/tools/Common/JitInterface/CorInfoInstructionSet.cs
… src/coreclr/src/tools/Common/JitInterface/ThunkGenerator/InstructionSetDesc.txt
… src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/X86/X86Base.PlatformNotSupported.cs
… src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/X86/X86Base.cs
… src/libraries/System.Utf8String.Experimental/src/System/Runtime/Intrinsics/Intrinsics.Shims.cs