Testing the assertion @tannergooding made in https://github.com/dotnet/runtime/issues/953#issuecomment-608949008 that adding x86 intrinsics is as easy as 1,2,3,4,5,6,7 😄

I wanted to get feedback on the introduction of an X86Base ISA to mirror what we have on ARM, and an `X86Base` class in S.R.I to house the methods for new instructions.  I recall @jkotas mentioning the possibility of exposing some of the x86 base instructions as internal HWIntrinsics that could be used from higher-level methods, although I can't seem to find the discussion now.

I've added HWIntrinsics implementations of `BSF` and `BSR` and used them as fallbacks in some of the `BitOperations` methods that use `LZCNT` and `TZCNT`.  Since BMI1 requires VEX encoding, it won't be supported on older processors or newer non-VEX processors like the Intel Atom line.  The fallbacks will also improve codegen in R2R images, where VEX encoding is disabled.

Benchmark results with BMI1 and LZCNT disabled:

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Xeon CPU E3-1505M v6 3.00GHz, 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.1.300-preview-015048
  [Host]     : .NET Core 5.0.0 (CoreCLR 5.0.20.16006, CoreFX 5.0.20.16006), X64 RyuJIT
  Job-XPOYMR : .NET Core 5.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), X64 RyuJIT
  Job-ROWHAK : .NET Core 5.0 (CoreCLR 42.42.42.42424, CoreFX 42.42.42.42424), X64 RyuJIT

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:DebugType=portable  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                  Method |                    Toolchain |       Mean |     Error |    StdDev |     Median |        Min |        Max | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |----------------------------- |-----------:|----------:|----------:|-----------:|-----------:|-----------:|------:|--------:|------:|------:|------:|----------:|
|   LeadingZeroCount_uint | \corerun-bsf-bsr\CoreRun.exe | 1,176.0 ns |   8.28 ns |   7.74 ns | 1,176.4 ns | 1,161.6 ns | 1,188.7 ns |  0.30 |    0.01 |     - |     - |     - |         - |
|   LeadingZeroCount_uint |  \corerun-master\CoreRun.exe | 3,867.1 ns | 172.27 ns | 198.39 ns | 3,855.8 ns | 3,509.7 ns | 4,214.0 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |           |           |            |            |            |       |         |       |       |       |           |
|  LeadingZeroCount_ulong | \corerun-bsf-bsr\CoreRun.exe | 1,156.3 ns |  22.38 ns |  23.94 ns | 1,158.5 ns | 1,115.4 ns | 1,204.8 ns |  0.36 |    0.01 |     - |     - |     - |         - |
|  LeadingZeroCount_ulong |  \corerun-master\CoreRun.exe | 3,185.6 ns |  73.96 ns |  82.21 ns | 3,178.5 ns | 3,091.1 ns | 3,362.5 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |           |           |            |            |            |       |         |       |       |       |           |
|               Log2_uint | \corerun-bsf-bsr\CoreRun.exe |   883.6 ns |   9.93 ns |   9.29 ns |   880.6 ns |   871.2 ns |   903.2 ns |  0.31 |    0.01 |     - |     - |     - |         - |
|               Log2_uint |  \corerun-master\CoreRun.exe | 2,806.4 ns |  44.82 ns |  41.93 ns | 2,807.6 ns | 2,730.5 ns | 2,872.6 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |           |           |            |            |            |       |         |       |       |       |           |
|              Log2_ulong | \corerun-bsf-bsr\CoreRun.exe |   875.5 ns |  12.38 ns |  11.58 ns |   876.3 ns |   856.2 ns |   896.4 ns |  0.22 |    0.00 |     - |     - |     - |         - |
|              Log2_ulong |  \corerun-master\CoreRun.exe | 4,058.1 ns |  42.23 ns |  39.51 ns | 4,059.0 ns | 4,001.2 ns | 4,138.1 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |           |           |            |            |            |       |         |       |       |       |           |
|  TrailingZeroCount_uint | \corerun-bsf-bsr\CoreRun.exe |   885.0 ns |   8.74 ns |   7.30 ns |   885.3 ns |   874.4 ns |   897.3 ns |  0.52 |    0.08 |     - |     - |     - |         - |
|  TrailingZeroCount_uint |  \corerun-master\CoreRun.exe | 1,710.8 ns | 206.28 ns | 237.55 ns | 1,607.4 ns | 1,509.7 ns | 2,174.5 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |           |           |            |            |            |       |         |       |       |       |           |
| TrailingZeroCount_ulong | \corerun-bsf-bsr\CoreRun.exe | 1,030.8 ns |  19.25 ns |  19.76 ns | 1,036.3 ns |   984.6 ns | 1,061.7 ns |  0.69 |    0.02 |     - |     - |     - |         - |
| TrailingZeroCount_ulong |  \corerun-master\CoreRun.exe | 1,484.5 ns |  45.58 ns |  52.49 ns | 1,470.6 ns | 1,425.7 ns | 1,625.9 ns |  1.00 |    0.00 |     - |     - |     - |         - |

And enabled:

|                  Method |                    Toolchain |       Mean |    Error |   StdDev |     Median |        Min |        Max | Ratio | RatioSD | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------ |----------------------------- |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------:|--------:|------:|------:|------:|----------:|
|   LeadingZeroCount_uint | \corerun-bsf-bsr\CoreRun.exe |   632.8 ns |  8.39 ns |  7.85 ns |   630.3 ns |   624.2 ns |   649.7 ns |  0.99 |    0.02 |     - |     - |     - |         - |
|   LeadingZeroCount_uint |  \corerun-master\CoreRun.exe |   636.9 ns |  8.66 ns |  8.10 ns |   636.2 ns |   623.9 ns |   653.7 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |          |          |            |            |            |       |         |       |       |       |           |
|  LeadingZeroCount_ulong | \corerun-bsf-bsr\CoreRun.exe |   634.0 ns |  8.36 ns |  7.82 ns |   632.2 ns |   624.7 ns |   649.9 ns |  1.00 |    0.01 |     - |     - |     - |         - |
|  LeadingZeroCount_ulong |  \corerun-master\CoreRun.exe |   631.2 ns |  8.40 ns |  7.86 ns |   629.3 ns |   616.1 ns |   643.2 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |          |          |            |            |            |       |         |       |       |       |           |
|               Log2_uint | \corerun-bsf-bsr\CoreRun.exe |   729.5 ns | 14.35 ns | 12.72 ns |   728.4 ns |   711.4 ns |   751.3 ns |  0.70 |    0.05 |     - |     - |     - |         - |
|               Log2_uint |  \corerun-master\CoreRun.exe | 1,021.8 ns | 62.09 ns | 71.51 ns | 1,005.0 ns |   926.7 ns | 1,175.7 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |          |          |            |            |            |       |         |       |       |       |           |
|              Log2_ulong | \corerun-bsf-bsr\CoreRun.exe |   735.5 ns |  7.79 ns |  7.29 ns |   737.1 ns |   720.3 ns |   748.9 ns |  0.42 |    0.01 |     - |     - |     - |         - |
|              Log2_ulong |  \corerun-master\CoreRun.exe | 1,743.7 ns | 33.28 ns | 34.18 ns | 1,739.0 ns | 1,684.3 ns | 1,821.5 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |          |          |            |            |            |       |         |       |       |       |           |
|  TrailingZeroCount_uint | \corerun-bsf-bsr\CoreRun.exe |   626.5 ns |  9.07 ns |  8.48 ns |   625.5 ns |   614.8 ns |   640.9 ns |  0.99 |    0.02 |     - |     - |     - |         - |
|  TrailingZeroCount_uint |  \corerun-master\CoreRun.exe |   633.5 ns |  8.48 ns |  7.93 ns |   631.2 ns |   622.1 ns |   648.7 ns |  1.00 |    0.00 |     - |     - |     - |         - |
|                         |                              |            |          |          |            |            |            |       |         |       |       |       |           |
| TrailingZeroCount_ulong | \corerun-bsf-bsr\CoreRun.exe |   636.7 ns | 11.54 ns | 10.79 ns |   632.3 ns |   621.0 ns |   661.9 ns |  1.01 |    0.02 |     - |     - |     - |         - |
| TrailingZeroCount_ulong |  \corerun-master\CoreRun.exe |   630.5 ns |  7.80 ns |  7.29 ns |   627.5 ns |   620.2 ns |   648.0 ns |  1.00 |    0.00 |     - |     - |     - |         - |

Note the nice little improvement to `Log2` by replacing a subtraction with xor to reverse the index returned by `LZCNT`