@tannergooding 2020-04-07
CC. @CarolEidt and @echesakovMSFT 

Also CC. @davidwrighton to help validate the new `IsSupported` check is hooked up correctly