@saucecontrol 2020-04-07
> Looks like something isn't quite hooked up right as `X86Base.IsSupported` is returning false for various tests.

I saw that `ArmBase` is explicitly added with the , even though it looks like it should already be there from the VM side:

https://github.com/dotnet/runtime/blob/master/src/coreclr/src/jit/compiler.cpp#L2313-L2320

https://github.com/dotnet/runtime/blob/master/src/coreclr/src/vm/codeman.cpp#L1485-L1487

I couldn't see any cases where X86Base wasn't working, but maybe it needs the same treatment if you're seeing a failure.