@tannergooding 2020-04-07
> I couldn't see any cases where X86Base wasn't working, but maybe it needs the same treatment if you're seeing a failure.

Seeing failures in CI for x86.