@saucecontrol 2020-04-07
Ah, I found one:

```
System.PlatformNotSupportedException : Operation is not supported on this platform.
      Stack Trace:
        /_/src/libraries/System.Private.CoreLib/src/System/Runtime/Intrinsics/X86/X86Base.cs(34,0): at System.Runtime.Intrinsics.X86.X86Base.X64.BitScanReverse(UInt64 value)
```

So it looks like `X86Base.X64.IsSupported` is returning true when it shouldn't.  I'll investigate