@tannergooding 2020-04-07
We need the instruction sets added here: https://github.com/dotnet/runtime/blob/master/src/coreclr/src/jit/compiler.cpp#L2205, in the same way as we do for ARM here: https://github.com/dotnet/runtime/blob/master/src/coreclr/src/jit/compiler.cpp#L2316

It looks like all the other JIT bits exist (comparing to `ArmBase` and `Sse.X64`)