@saucecontrol 2020-04-07
The way I read it, the explicit addition of ArmBase is redundant because it's passed in from the VM side in the same way the higher-level ISAs are.  Since I've added X86Base unconditionally here:

https://github.com/dotnet/runtime/blob/47bc99ba0b45f8cf2144e590013c61ad911865db/src/coreclr/src/vm/codeman.cpp#L1302

It's always available unless removed based on the config knobs, just like SSE+ are.

Am I missing something there?