@tannergooding 2020-04-07
> The way I read it, the explicit addition of ArmBase is redundant because it's passed in from the VM side in the same way the higher-level ISAs are. Since I've added X86Base unconditionally here:

@davidwrighton, did the base checks get inverted? That is should https://github.com/dotnet/runtime/blob/master/src/coreclr/src/jit/compiler.cpp#L2205 be:
```
if (!JitConfig.EnableHWIntrinsic())
{
    // Dummy ISAs for simplifying the JIT code
    instructionSetFlags.RemoveInstructionSet(InstructionSet_Vector128);
    instructionSetFlags.RemoveInstructionSet(InstructionSet_Vector256);
}
```

You are meant to be able to disable the "baseline" as well, so you can test your software fallback.