@saucecontrol 2020-04-07
The Vector ISAs don't actually exist, so they're not passed into the JIT.  Unless `JitConfig.EnableHWIntrinsic` is true, they never get added so you get the software fallbacks.

The real ISAs work the opposite way.  The VM passes them in if the hardware supports them and the JIT disables them if the config says to.

`X86Base` is working as expected with that model.  It reports `IsSupported` if COMPlus_EnableHWIntrinsic is on and not if it's not.