@saucecontrol 2020-04-07
It's probably a good idea to make `ArmBase` and `X86Base` handled consistently in compiler.cpp.  I actually found the explicit addition of `ArmBase` with the dummy ISAs confusing, so I'd be in favor of removing that if it's not actually necessary.

If I've misunderstood any of that, I guess we should add `X86Base` explicitly to make them match 🤷‍♂