@davidwrighton 2020-04-07
@tannergooding @saucecontrol Vector64/128/256 are set by the JIT itself as Tanner noted. 

One of the changes I did not realize I made was to make ArmBase set by the VM side. One of the issues is that the VM now attempts to pass a rationalized set of instruction set values to the JIT, and not enabling ArmBase, then requires the VM side to disable all of the other instruction sets. We could tweak that if we needed to but I need to know its necessary before I futz with it.

I would avoid marking the X86Base as being implied by support for SSE, as that means if it is disabled, then SSE will be disabled.