@saucecontrol 2020-04-07
> I would avoid marking the X86Base as being implied by support for SSE, as that means if it is disabled, then SSE will be disabled.

There's presently no way to disable X86Base other than to disable HWIntrisics entirely.  Isn't that the same way ArmBase works?  Or would this be a case where disabling HWIntrinsics also disables SIMD?

> And yes, the current model doesn't make sense where we explicitly set the ArmBase instruction set in both the jit and the VM side of the fence.

Would removing it from the JIT side be ok for now, given the larger changes necessary to separate it on the VM side?