@tannergooding 2020-04-07
> There's presently no way to disable X86Base other than to disable HWIntrinsics entirely. Isn't that the same way ArmBase works? Or would this be a case where disabling HWIntrinsics also disables SIMD?

That is the expectation.

The helper types and the "base" ISAs don't have their own `COMPlus` knob, instead they are controlled by the `COMPlus_EnableHWIntrinsics` knob.
They still participate in the hierarchy and so doing `EnableHWIntrinsics=0` will also disable `Sse` and others (we can't very well use `Sse` if the user has disabled `HWIntrinsics` in general, after all).