@tannergooding 2020-04-07
> I would avoid marking the X86Base as being implied by support for SSE, as that means if it is disabled, then SSE will be disabled.

That is exactly what we want. Disabling `X86Base` requires disabling HWIntrinsics in general, which should also disable `SSE` (only intrinsics, not general JIT support outside HWIntrinsics of course).