@davidwrighton 2020-04-08
Ah, ok then. Well, then the model should be ArmBase, and X86Base are set unconditionally by the VM side of the world. And if HWIntrinsics aren't enabled, then those two should be explicitly disabled by the JIT. @saucecontrol, would you like to do that as part of this PR?