@saucecontrol 2020-04-08
> would you like to do that as part of this PR?

Sure, I don't mind cleaning that up here if I can get the model straight.

In my first attempt at this code, I actually did explicitly disable X86Base when `JitConfig.EnableHWIntrinsic` was false, but that caused some asserts in the JIT SIMD code when running with `COMPlus_EnableHWIntrinsic=0`.  IIRC, the assert was being hit on `Vector<T>.Create()`.  It seems like this might have been a side effect of SSE+ being implicitly disabled with X86Base.  I found that by leaving X86Base enabled, everything just magically worked, which I assume is because the `IsSupported` checks returned false with HWIntrinsics off but the higher ISAs were still available in the SIMD context.

The interaction between the HWIntrinsic and SIMD logic in the JIT and their dependence on the ISA checks is still a bit of a mystery to me, so I'm not sure how to resolve that... unless it's simply a matter of not making SSE depend on X86Base.  I don't know if the same would apply on the ARM side since I don't know how much of the SIMD code is wired up for ARM.